Class Separated:Separate Class for Audio Recording and Audio Playing
