+++ app/src/main/java/com/example/keshavdulal/a14_simple_drawing/AudioPlayClass.java
@@ -1,7 +1,61 @@
 package com.example.keshavdulal.a14_simple_drawing;
 
+import android.graphics.Color;
+import android.media.AudioFormat;
+import android.media.AudioManager;
+import android.media.AudioTrack;
+import android.os.Environment;
+
+import java.io.BufferedInputStream;
+import java.io.DataInputStream;
+import java.io.File;
+import java.io.FileInputStream;
+import java.io.FileNotFoundException;
+import java.io.IOException;
+import java.io.InputStream;
+
 /**
  * Created by amitgupta on 7/5/2016.
  */
 public class AudioPlayClass {
-}
+
+    //Start of playRecord()
+    public  static  void playRecord(){
+
+        File filePcm = new File(Environment.getExternalStorageDirectory(), "Sound.pcm");
+        int shortSizeInBytes = Short.SIZE/Byte.SIZE;
+
+        int bufferSizeInBytes = (int)(filePcm.length()/shortSizeInBytes);
+        short[] audioData = new short[bufferSizeInBytes];
+
+        try {
+            InputStream inputStream = new FileInputStream(filePcm);
+            BufferedInputStream bufferedInputStream = new BufferedInputStream(inputStream);
+            DataInputStream dataInputStream = new DataInputStream(bufferedInputStream);
+            int i = 0;
+            while(dataInputStream.available() > 0){
+                audioData[i] = dataInputStream.readShort();
+                i++;
+            }
+
+            dataInputStream.close();
+
+            AudioTrack audioTrack = new AudioTrack(
+                    AudioManager.STREAM_MUSIC,
+                    44100,
+                    AudioFormat.CHANNEL_IN_MONO,
+                    AudioFormat.ENCODING_PCM_16BIT,
+                    bufferSizeInBytes,
+                    AudioTrack.MODE_STREAM);
+
+            audioTrack.play();
+            audioTrack.write(audioData, 0, bufferSizeInBytes);
+
+        } catch (FileNotFoundException e) {
+            e.printStackTrace();
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+    }//End of playRecord()
+
+}//END
